moved endpoints from DELETE /[some-route]/delete to DELETE /[some-route]
